requirements.txt
requests==2.31.0
pandas==2.2.1
aiohttp==3.9.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
tqdm==4.66.2 